+++ addon/services/objectlistview-events.js
@@ -5,6 +5,7 @@
 import Service from '@ember/service';
 import Evented from '@ember/object/evented';
 import EmberMap from '@ember/map';
+import { computed } from '@ember/object';
 import { isNone } from '@ember/utils';
 import { deprecatingAlias } from '@ember/object/computed';
 import { inject as service } from '@ember/service';
@@ -99,6 +100,16 @@ export default Service.extend(Evented, {
   */
   refreshListTrigger(componentName) {
     this.trigger('refreshList', componentName);
+  },
+
+  /**
+    Trigger for "refresh list" event in OLV component by name.
+
+    @method refreshListOnlyTrigger
+    @param {String} componentName The name of OLV component.
+  */
+  refreshListOnlyTrigger(componentName) {
+    this.trigger('refreshListOnly', componentName);
   },
 
   /**
@@ -208,6 +219,19 @@ export default Service.extend(Evented, {
   },
 
   /**
+    Trigger for "setSorting" event in route.
+    Event name: setSorting.
+
+    @method setSortingTrigger
+
+    @param {String} componentName The name of object-list-view component.
+    @param {Array} sorting Array of sorting definitions.
+  */
+  setSortingTrigger(componentName, sorting = []) {
+    this.trigger('setSorting', componentName, sorting);
+  },
+
+  /**
     Trigger for "geSortApply" event in object-list-view.
     Event name: geSortApply.
 
@@ -265,13 +289,13 @@ export default Service.extend(Evented, {
   },
 
   /**
-    Current limit function for OLV.
+    Current limit functions for OLV by componentNames.
 
-    @property currentLimitFunction
-    @type BasePredicate
-    @default undefined
+    @property currentLimitFunctions
+    @type Object
+    @default {}
   */
-  currentLimitFunction: undefined,
+  currentLimitFunctions: computed(() => { return {}; }).readOnly(),
 
   /**
     Form's loading state.
@@ -295,19 +319,21 @@ export default Service.extend(Evented, {
     @method setLimitFunction
 
     @param {BasePredicate} limitFunction Current limit function.
+    @param {String} componentName Component name.
   */
-  setLimitFunction(limitFunction) {
-    this.set('currentLimitFunction', limitFunction instanceof BasePredicate ? limitFunction : undefined);
+  setLimitFunction(limitFunction, componentName) {
+    this.set(`currentLimitFunctions.${componentName}`, limitFunction instanceof BasePredicate ? limitFunction : undefined);
   },
 
   /**
     Gets current limit function for OLV.
 
     @method getLimitFunction
+    @param {String} componentName Component name.
     @return {BasePredicate} Current limit function.
   */
-  getLimitFunction() {
-    return this.get('currentLimitFunction');
+  getLimitFunction(componentName) {
+    return this.get(`currentLimitFunctions.${componentName}`);
   },
 
   /**